first review problems fixed

+++ lesson_1/triangle_area.rb
@@ -1,7 +1,7 @@
 print "Введите основание треугольника: "
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 print "Введите высоту треугольника: "
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 area = 0.5 * base * height
